requirements.txt
argparse
feedparser
pyyaml
SQLAlchemy
requests
Jinja2